pyproject.toml
[tool.poetry]
name = "approver-backend"
version = "0.1.0"
description = ""
authors = ["readyyyk <mr.sairexer@gmail.com>", "halone228 <haloneprog@gmail.com>"]
readme = "README.md"


[tool.poetry.scripts]
start = "main:start"
dev = "main:dev"
test = "main:test"


[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.25"
fastapi = "^0.109.0"
pydantic = "^2.5.3"
cryptography = "^42.0.1"
bcrypt = "==4.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
uvicorn = "^0.27.0"
loguru = "^0.7.2"


[tool.poetry.group.dev.dependencies]
aiosqlite = "^0.19.0"
python-dotenv = "^1.0.1"
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
randomuser = "^1.6"
pytest-dependency = "^0.6.0"
httpx = "^0.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
